requirements.txt
aiohttp>=3.9.0
discord.py>=2.3.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
tzdata>=2024.1
